refactor: move code to files
